move some types to AsymmetricCryptoFacade that belong there
